added different settings to calibration_node.cpp; changed flags for calibratecamera !! need to check; tvec = -R*tvec !!!! need to check;

+++ calibration/src/calibration_node.cpp
@@ -1,6 +1,8 @@
-//
-// Created by nico on 18.06.21.
-//
+// cameraCalibration: rotation matrix is pretty close to provided transofmrations,
+// translation vector is very different
+// TODO: fix cameraCalibration translation vector
+
+
 #include <iostream>
 #include <string>
 #include <sstream>
@@ -13,14 +15,14 @@
 #include <cv_bridge/cv_bridge.h>
 #include <opencv2/imgcodecs.hpp>
 #include "cv_handeye.h"
-#include <geometry_msgs/TransformStamped.h>
-#include <tf/transform_listener.h>
+//#include <geometry_msgs/TransformStamped.h>
+//#include <tf/transform_listener.h>
 #include <boost/filesystem.hpp>
 #include <eigen3/Eigen/Dense>
 
 #include <opencv2/core/mat.hpp>
 
-#include <tf2_ros/transform_listener.h>
+//#include <tf2_ros/transform_listener.h>
 #include "calibration/handEyeCalibration.h"
 
 
@@ -28,14 +30,28 @@ using namespace cv;
 using namespace std;
 using namespace Eigen;
 
-// define number of frames  to be extracted and used for calibration... later on this will be number of good frame pairs
+//// settings:
+// define number of frames to be extracted and used for calibration:
 int n_frames = 10;
 
-// package path
-string path = "/home/lars/cal_data";
+// only one of the following should be true
+
+// set true to write joint states and images to files:
+bool write_files = false;
+// set true to receive frames + joint states from get_images_node, and calibrate from those:
+bool receive_frames = false; // <--- standard setting
+// true to read from provided txt files:
+bool from_folder = false;
+// use written images for cameracal and joint_states.txt for handeye:
+bool from_folder_ = true;
+// using given k and K matrices:
+bool use_preset = false;
+// calibration data path:
+string path = "/home/nico/cal_data";
 //string path = "/home/rnm_grp1/rgb";
 
-// counters for the filenames
+
+// counters for filenames
 int pose_rgb = 1;
 int pose_ir = 1;
 int js_count = 0;
@@ -47,23 +63,10 @@ std::vector<cv::Mat> cameraPosesR;
 std::vector<cv::Mat> R_target2cam;
 std::vector<cv::Mat> t_target2cam;
 
-
-
-// container for joint states as geometry_msgs
-//vector<tf::StampedTransform> allJointStates;
 vector<vector<double>> allJointStates;
 vector<array<cv::Mat,2>> allTransforms;
 
-
-
-
-// tf listener to get robot pose
-tf::TransformListener* tfListener;
-//tf2_ros::TransformListener tfListener;
-cv::String robot_base = "panda_link0";
-cv::String gripper = "panda_link8";
-
-// from forward_kinematics:
+// helper function from forward_kinematics:
 MatrixXd get_transformationmatrix(const float theta,
                                   const float a,
                                   const float d,
@@ -75,13 +78,14 @@ MatrixXd get_transformationmatrix(const float theta,
   return ret_mat;
 }
 
-// forward_kinematics:
+// calculating transformation (pose) from joint angles,
+// adapted version from forward_kinematics
 array<Mat, 2> get_forward_kinematics_transformation(VectorXd a, vector<double> joint_angles_,
                                                       VectorXd d, VectorXd alpha) {
   array<MatrixXd, 8> a_;
   Matrix4d A_total;
-  Mat R_gripper2base;
-  Mat t_gripper2base;
+  Mat one_R_gripper2base = Mat::zeros(3,3, CV_64F);
+  Mat one_t_gripper2base = Mat::zeros(3,1,CV_64F);
 
   for (int i = 0; i < 7; i++) {
     a_.at(i) = get_transformationmatrix(joint_angles_[i], a(i), d(i), alpha(i));
@@ -91,19 +95,18 @@ array<Mat, 2> get_forward_kinematics_transformation(VectorXd a, vector<double> j
 
   for (int i = 0; i < A_total.rows() - 1; i++) {
     for (int j = 0; j < A_total.cols() - 1; j++) {
-      R_gripper2base.at<double>(i, j) = A_total(i, j);
+      one_R_gripper2base.at<double>(i, j) = A_total(i, j);
     }
-    t_gripper2base.at<double>(i) = A_total(i, 3);
+    one_t_gripper2base.at<double>(i) = A_total(i, 3);
   }
 
-  return {R_gripper2base, t_gripper2base};
+  return {one_R_gripper2base, one_t_gripper2base};
 }
 
 // saving jointState msg
 void jointStatesWrite(const sensor_msgs::JointState& msg) {
+  //// get joint angles
   if(js_count<n_frames) {
-    // get robot pose
-    // TODO: Rotation matrix from msg + endeffector pos from msg
     //tf::StampedTransform joint_state;
     //tfListener->lookupTransform(robot_base, gripper, ros::Time(0), joint_state);
 
@@ -111,7 +114,32 @@ void jointStatesWrite(const sensor_msgs::JointState& msg) {
     one_joint_states = msg.position;
     allJointStates.push_back(one_joint_states);
     js_count++;
+
+
+    cout << "one_joint_states: " << endl;
+    cout << "[";
+    for (int i=0; i < one_joint_states.size(); i++) {
+      cout << one_joint_states[i] << ", ";
+    }
+    cout << "]" << endl;
   }
+}
+void jointStatesWriteFile(const sensor_msgs::JointState& msg) {
+  string co = ",";
+  ofstream foutput;
+  ifstream finput;
+  finput.open(path + "/joint_states.txt");
+  foutput.open(path + "/joint_states.txt", ios::app);
+
+  vector<double, allocator<void>::rebind<double>::other> position = msg.position;
+  if (finput.is_open()) {
+    foutput << position[0] << co << position[1] << co << position[2] << co << position[3]
+            << co << position[4] << co << position[5] << co << position[6] << co
+            << endl;
+    cout << "joint_states written to '" << path << "/joint_states.txt" << endl;
+  }
+  finput.close();
+  foutput.close();
 }
 
 // ir image writing
@@ -123,7 +151,7 @@ void irImageWrite(const sensor_msgs::ImageConstPtr& msg) {
 
     // create filename
     stringstream ss;
-    string name = "/pose_";
+    string name = "/imgs/pose_";
     string type = ".jpg";
     ss << path << name << (pose_ir) << type; // pose = counter, set at start
     string filename = ss.str();
@@ -166,23 +194,22 @@ void rgbImageWrite(const sensor_msgs::ImageConstPtr& msg) {
 
 // camera calibration
 int cameraCalibration() {
-  /* TODO: make sure joint states are synced with images*/
   cv::Size rgbFrameSize(2048, 1536);
-  cv::Size irFrameSize(640, 576);
+  //cv::Size irFrameSize(640, 576);
 
   std::vector<cv::String> rgbFileNames(n_frames);
-  for(int i = 0; i<n_frames; i++){
-    rgbFileNames[i] = "/home/lars/cal_data/imgs/pose_" + to_string(i+1) + ".jpg";
+  for (int i = 0; i < n_frames; i++) {
+    rgbFileNames[i] = "/home/nico/cal_data/imgs/pose_" + to_string(i + 1) + ".jpg";
   }
-  //std::vector<cv::String> irFileNames;
-  // std::string rgbFolder("/home/lars/CLionProjects/CameraCalibrationtests/cal_imgs/rgb/*.jpg");
-  //std::string rgbFolder("/home/nico/catkin_ws/src/frame_reader/cal_imgs/new/rgb/*.jpg");
-  //cv::glob(rgbFolder, rgbFileNames, true);  // load rgb images into opencv
-  //std::string irFolder("/home/nico/catkin_ws/src/frame_reader/cal_imgs/new/ir/*.jpg");
-  // std::string irFolder("/home/lars/CLionProjects/CameraCalibrationtests/cal_imgs/ir/*.jpg");
-  //cv::glob(irFolder, irFileNames, true);
+  // std::vector<cv::String> irFileNames;
+  //  std::string rgbFolder("/home/lars/CLionProjects/CameraCalibrationtests/cal_imgs/rgb/*.jpg");
+  // std::string rgbFolder("/home/nico/catkin_ws/src/frame_reader/cal_imgs/new/rgb/*.jpg");
+  // cv::glob(rgbFolder, rgbFileNames, true);  // load rgb images into opencv
+  // std::string irFolder("/home/nico/catkin_ws/src/frame_reader/cal_imgs/new/ir/*.jpg");
+  //  std::string irFolder("/home/lars/CLionProjects/CameraCalibrationtests/cal_imgs/ir/*.jpg");
+  // cv::glob(irFolder, irFileNames, true);
 
-  //std::vector<std::vector<cv::Point3f>> irObjP;   // Checkerboard world coordinates
+  // std::vector<std::vector<cv::Point3f>> irObjP;   // Checkerboard world coordinates
   std::vector<std::vector<cv::Point3f>> rgbObjP;  // Checkerboard world coordinates
 
   // Define checkerboard Parameters for both frames
@@ -191,9 +218,9 @@ int cameraCalibration() {
   cv::Size patternSize(9 - 1, 6 - 1);  // Number of inner corners per a chessboard row and column
 
   std::vector<std::vector<cv::Point2f>> rgbcorners(rgbFileNames.size());  // corners in rgb frames
-  //std::vector<std::vector<cv::Point2f>> ircorners(irFileNames.size());    // corners in ir frames
+  // std::vector<std::vector<cv::Point2f>> ircorners(irFileNames.size());    // corners in ir frames
   std::vector<cv::Point2f> rgbimgp;  // Define the rgb img point
-  //std::vector<cv::Point2f> irimgp;   // Define the ir img point
+  // std::vector<cv::Point2f> irimgp;   // Define the ir img point
 
   // Define the world coordinates for 3D points
 
@@ -224,13 +251,13 @@ int cameraCalibration() {
     cv::cvtColor(rgbimg, rgbgray, cv::COLOR_RGB2GRAY);
     bool rgbPatternFound = cv::findChessboardCorners(rgbgray, patternSize, rgbcorners[i2],
                                                      cv::CALIB_CB_ADAPTIVE_THRESH
-                                                     //+ cv::CALIB_CB_NORMALIZE_IMAGE
-                                                     + cv::CALIB_CB_FILTER_QUADS);
+                                                         //+ cv::CALIB_CB_NORMALIZE_IMAGE
+                                                         + cv::CALIB_CB_FILTER_QUADS);
     // Use cv::cornerSubPix() to refine the found corner detections with default values given by opencv
     if (rgbPatternFound) {
       vector<Point_<float>> temp_ = rgbcorners[i2];
       cv::cornerSubPix(
-          rgbgray, temp_, cv::Size(11, 11), cv::Size(-1, -1),// winsize 11,11 gute
+          rgbgray, temp_, cv::Size(11, 11), cv::Size(-1, -1),  // winsize 11,11 gute
           cv::TermCriteria(cv::TermCriteria::EPS + cv::TermCriteria::MAX_ITER, 100, 0.01));
       rgbcorners[i2] = temp_;
 
@@ -238,11 +265,9 @@ int cameraCalibration() {
       i2++;
     } else {
       rgbcorners.erase(next(rgbcorners.begin(), i - i_deleted));
-      cout << "!!!!!!!!File " << rgbFileNames[i] << "not used!!!!!!!!" << endl;
-      allJointStates.erase(next(allJointStates.begin(), i - i_deleted));
-      cout << "!!!!!!!!Joint state " << rgbFileNames[i] << "not used!!!!!!!!" << endl;
       rgbFileNames.erase(next(rgbFileNames.begin(), i - i_deleted));
-      cout << "!!!!!!!!Joint state " << rgbFileNames[i] << "not used!!!!!!!!" << endl;
+      allJointStates.erase(next(allJointStates.begin(), i - i_deleted));
+      cout << "!!!!!!!!File " << rgbFileNames[i] << "not used!!!!!!!!" << endl;
       i_deleted++;
     }
 
@@ -250,12 +275,12 @@ int cameraCalibration() {
   }
   cout << "All RGB Corners detected and safed in rgbcorners\n";
 
-  for(int i =0; i <  rgbFileNames.size() ; i++){
+  for (int i = 0; i < rgbFileNames.size(); i++) {
     // Display the detected pattern on the chessboard
     rgbimg = cv::imread(rgbFileNames[i]);
     cv::drawChessboardCorners(rgbimg, patternSize, rgbcorners[i], true);
-    //cv::imshow("RGB chessboard corner detection", rgbimg);
-    //cv::waitKey(0);
+    // cv::imshow("RGB chessboard corner detection", rgbimg);
+    // cv::waitKey(0);
   }
 
   /*
@@ -294,32 +319,49 @@ int cameraCalibration() {
   cout << "All IR Corners detected and safed in ircorners\n";
    */
 
-  // Calibrate the rgb frame intrinsics
-  cv::Mat rgbK;  // calibration Matrix K
-  Mat rgbk = (Mat1d(1, 8));
-  std::vector<cv::Mat> rvecs, tvecs;
-  std::vector<double> stdIntrinsics, stdExtrinsics, perViewErrors;
-  int flags = cv::CALIB_FIX_S1_S2_S3_S4 + cv::CALIB_FIX_TAUX_TAUY +
-              cv::CALIB_RATIONAL_MODEL;  // used to compute k4-k6
+    // Calibrate the rgb frame intrinsics
+    Mat rgbK;  // calibration Matrix K
+    Mat rgbk = (Mat1d(1, 8));
+    std::vector<cv::Mat> rvecs, tvecs;
+    std::vector<double> stdIntrinsics, stdExtrinsics, perViewErrors;
+    //int flags = cv::CALIB_FIX_S1_S2_S3_S4 + cv::CALIB_FIX_TAUX_TAUY +
+                cv::CALIB_RATIONAL_MODEL;  // used to compute k4-k6
+    int flags = cv::CALIB_FIX_ASPECT_RATIO + cv::CALIB_RATIONAL_MODEL;
+    // try these cv::CALIB_ZERO_TANGENT_DIST + cv::CALIB_FIX_PRINCIPAL_POINT
 
-  std::cout << "Calibrating rgb intrinsics...\n" << endl;
+    std::cout << "Calibrating rgb intrinsics...\n" << endl;
 
-  float rgberror =
-      cv::calibrateCamera(rgbObjP, rgbcorners, rgbFrameSize, rgbK, rgbk, rvecs, tvecs, flags);
-  std::cout << "Reprojection error of rgb frames = " << rgberror << "\nK =\n"
-            << rgbK << "\nk=\n"
-            << rgbk << std::endl;
+    float rgberror =
+        cv::calibrateCamera(rgbObjP, rgbcorners, rgbFrameSize, rgbK, rgbk, rvecs, tvecs, flags);
+    std::cout << "Reprojection error of rgb frames = " << rgberror << "\nK =\n"
+              << rgbK << "\nk=\n"
+              << rgbk << std::endl;
 
   cv::Mat rvec;
   cv::Mat tvec;
   std::cout << "rgbObjP.size: " << rgbObjP.size() << " rgbcorners.size: " << rgbcorners.size() << " rgbFileNames.size: " << rgbFileNames.size() << std::endl;
+
+
+  if (use_preset == 1) {
+    float rgbKData[9] = {971.997, 0, 1022.58, 0, 971.906, 775.268, 0, 0, 1};
+    float rgbkData[14] = {0.781141, -3.04307, 0.000614642, -2.35504e-05,
+                          1.7224,   0.656092, -2.86714,    1.64986,
+                          0,        0,        0,           0,
+                          0,        0};  // k1,k2,p1,p2,k3,k4,k5,k6,s1,s2,s3,s4,tx,ty
+
+    Mat rgbK = Mat(3, 3, CV_32F, rgbKData);
+    Mat rgbk = Mat(1, 14, CV_32F, rgbkData);
+    cout << "rgbk: " << rgbk << endl;
+    cout << "rgbK: " << rgbK << endl;
+  }
+
+
+
+
   for (int i = 0; i < rgbFileNames.size(); i++) {
     cv::InputArray rgb1ObjP = rgbObjP[i];
     cv::InputArray rgb1corners = rgbcorners[i];
     cv::solvePnPRansac(rgb1ObjP, rgb1corners, rgbK, rgbk, rvec, tvec);
-
-    cout << "rvec camera" << rvec << endl;
-    cout << "tvec camera" << tvec << endl;
 
     //Mat R;
     //Rodrigues(rvec, R);
@@ -331,8 +373,11 @@ int cameraCalibration() {
     cv::Rodrigues(rvec, temp);
     temp = temp.t();
     R_target2cam.push_back(temp);
-    t_target2cam.push_back(tvec);
     cout << "R_target2cam: " << temp << endl;
+
+    tvec = -temp*tvec;
+
+    t_target2cam.push_back(tvec);
     cout << "t_target2cam: " << tvec << endl;
   }
 
@@ -458,6 +503,88 @@ void transform2rv(tf::StampedTransform transform, cv::Mat& rvec, cv::Mat& tvec){
 }
 */
 
+
+void readJointStates() {
+  //// joint states
+  try {
+    ifstream file(path + "/joint_states.txt");
+    string line;
+    while (getline(file, line)) {
+      vector<double> one_joint_states;
+      istringstream stream(line);
+      double d;
+      while (stream >> d) {
+        one_joint_states.push_back(d);
+      }
+      /*
+      cout << "one_joint_states: " << endl;
+      cout << "[";
+      for (int i=0; i < one_joint_states.size(); i++) {
+        cout << one_joint_states[i] << ", ";
+      }
+      cout << "]" << endl;
+       */
+      allJointStates.push_back(one_joint_states);
+    }
+    cout << "allJointStates.size: " << allJointStates.size() << endl;
+    ROS_INFO("Done reading joint states.");
+    cout << endl;
+  }
+  catch (...) {
+    ROS_INFO("Error while reading joint states!!!");
+    cout << endl;
+  }
+}
+
+void readCheckPoses() {
+  //// checkerboard poses
+  // reading: 40 lines, each line consists of 16 entries (4x4 homogeneous rot + trans matrix)
+  // need to split into R matrix and t vector
+
+  // first pushing R entries of one matrix to oneR_target2cam, t equivalently
+  // then pushing those matrices to the vectors of matrices (R_target2cam, t_tar...)
+  try {
+    int rot_count = 1; // for each row, first 3 entries to R mat, 4th entry to t mat, repeat 3 times, then toss next four entries (0 0 0 1), 16 in total, then repeat for next line
+    ifstream file(path + "/trk.txt");
+    string line;
+    while (getline(file, line)) {
+      istringstream stream(line);
+      double d;
+      Mat oneR_target2cam;
+      Mat onet_target2cam;
+      while (stream >> d) {
+        if (rot_count == 4 || rot_count == 8 || rot_count == 12) {
+          onet_target2cam.push_back(d);
+          rot_count++;
+        }
+        else if (rot_count < 12){
+          oneR_target2cam.push_back(d);
+          rot_count++;
+        }
+        if (rot_count == 13) {
+          break;
+        }
+      }
+      oneR_target2cam = oneR_target2cam.reshape(1, 3);
+      cout << "oneR_target2cam: " << endl << oneR_target2cam << endl;
+      cout << "onet_target2cam: " << endl << onet_target2cam << endl;
+      t_target2cam.push_back(onet_target2cam);
+      R_target2cam.push_back(oneR_target2cam);
+      rot_count = 1;
+    }
+    cout << "R_target2cam.size: " << R_target2cam.size() << endl;
+    cout << "t_target2cam.size: " << t_target2cam.size() << endl;
+    ROS_INFO("Done reading checkerboard-to-camera-transformation.");
+    cout << endl;
+
+  }
+  catch (...) {
+    ROS_INFO("Error while reading checkerboard to camera transformation!!!");
+  }
+}
+
+
+
 cv::Mat R_cam2gripper, t_cam2gripper;
 // perform hand-eye calibration with the calculated transforms
 void handEye(){
@@ -550,23 +677,48 @@ bool handEyeOutput(calibration::handEyeCalibration::Request  &req,
 int main(int argc, char** argv) {
   ros::init(argc, argv,"calibration_node");
   ros::NodeHandle nh;
-
   image_transport::ImageTransport it(nh);
-  tfListener = new tf::TransformListener();
-  boost::filesystem::remove_all(path + "/imgs");
-  boost::filesystem::create_directory(path + "/imgs");
 
-  image_transport::Subscriber rgb_sub = it.subscribe("/calibration_rgb_img", 1, rgbImageWrite);
-  //image_transport::Subscriber ir_sub = it.subscribe("/calibration_ir_img", 1, irImageWrite);
-  ros::Subscriber joint_states_sub = nh.subscribe("/calibration_joint_states", 1, jointStatesWrite);
 
-  std::cout << "\nready to receive frames" << endl;
-  while(js_count < n_frames || rgb_count < n_frames) {
-    ros::spinOnce();
+  //tfListener = new tf::TransformListener();
+  if (write_files == true) {
+    boost::filesystem::remove(path + "/joint_states.txt");
+    boost::filesystem::remove_all(path + "/imgs");
+    boost::filesystem::create_directory(path + "/imgs");
+    image_transport::Subscriber rgb_sub = it.subscribe("/calibration_rgb_img", 1, rgbImageWrite);
+    ros::Subscriber joint_states_sub = nh.subscribe("/calibration_joint_states", 1, jointStatesWriteFile);
+    std::cout << "\nready to receive frames" << endl;
+    while (js_count < n_frames || rgb_count < n_frames) {
+      ros::spinOnce();
+    }
+    cout << "\ncollected " << js_count << " frames" << endl;
   }
-  cout << "\ncollected " << js_count << " frames" << endl;
-  cout << "starting camera calibration" << endl;
-  cameraCalibration();
+  if (receive_frames == true) {
+    boost::filesystem::remove(path + "/joint_states.txt");
+    boost::filesystem::remove_all(path + "/imgs");
+    boost::filesystem::create_directory(path + "/imgs");
+    image_transport::Subscriber rgb_sub = it.subscribe("/calibration_rgb_img", 1, rgbImageWrite);
+    //image_transport::Subscriber ir_sub = it.subscribe("/calibration_ir_img", 1, irImageWrite);
+    ros::Subscriber joint_states_sub = nh.subscribe("/calibration_joint_states", 1, jointStatesWrite);
+    std::cout << "\nready to receive frames" << endl;
+    while (js_count < n_frames || rgb_count < n_frames) {
+      ros::spinOnce();
+    }
+    cout << "\ncollected " << js_count << " frames" << endl;
+    cout << "starting camera calibration" << endl;
+    cameraCalibration();
+  }
+  if (from_folder == true) {
+    readJointStates();
+    // readEndPoses();
+    readCheckPoses();
+  }
+  if (from_folder_ == true) {
+    cout << "starting camera calibration" << endl;
+    readJointStates();
+    cameraCalibration();
+  }
+
   handEye();
 
   ros::ServiceServer service = nh.advertiseService("handEyeCalibration", handEyeOutput);
